refactor(Stream): again container + presenter, tests working with mocked stores
