+++ extension-mobx/src/stores/userStore.js
@@ -13,3 +13,4 @@ class UserStore {
 const userStore = new UserStore();
 
 export default userStore;
+export { UserStore };
